@@ -1,8 +1,9 @@
 package com.onzhou.tiktok;
 
 import android.content.Context;
-import android.support.v7.widget.AppCompatImageView;
 import android.util.AttributeSet;
+
+import androidx.appcompat.widget.AppCompatImageView;
 
 /**
  * @anchor: Andy
