@@ -1,11 +1,10 @@
 package com.onzhou.tiktok;
 
-import android.app.Activity;
-import android.graphics.Rect;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.onzhou.transition.TransitionParam;
 import com.onzhou.transition.TransitionUtils;
